+++ docs/src/examples/components/Menu/Variations/MenuExampleUnderlinedPrimary.shorthand.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+import { Menu } from 'stardust'
+
+const items = [
+  { key: 'editorials', content: 'Editorials' },
+  { key: 'review', content: 'Reviews' },
+  { key: 'events', content: 'Upcoming Events' },
+]
+
+class MenuExampleUnderlinedPrimaryShorthand extends React.Component {
+  render() {
+    return <Menu defaultActiveIndex={0} items={items} shape="underlined" type="primary" />
+  }
+}
+
+export default MenuExampleUnderlinedPrimaryShorthand