+++ docs/src/examples/components/Menu/index.tsx
@@ -1,9 +1,11 @@
 import React from 'react'
 import Types from './Types'
+import Variations from './Variations'
 
 const MenuExamples = () => (
   <div>
     <Types />
+    <Variations />
   </div>
 )