+++ src/components/Menu/MenuItem.tsx
@@ -6,11 +6,14 @@ import React from 'react'
 import { childrenExist, createShorthandFactory, customPropTypes, UIComponent } from '../../lib'
 
 import menuItemRules from './menuItemRules'
+import menuVariables from './menuVariables'
 
 class MenuItem extends UIComponent<any, any> {
   static displayName = 'MenuItem'
 
   static className = 'ui-menu__item'
+
+  static variables = menuVariables
 
   static create: Function
 
@@ -48,7 +51,9 @@ class MenuItem extends UIComponent<any, any> {
     pointing: PropTypes.bool,
 
     /** The menu can have primary or secondary type */
-    type: PropTypes.oneOf(['default', 'primary', 'secondary']),
+    type: PropTypes.oneOf(['primary', 'secondary']),
+
+    shape: PropTypes.oneOf(['pills', 'pointing', 'underlined']),
   }
 
   static defaultProps = {
@@ -64,6 +69,7 @@ class MenuItem extends UIComponent<any, any> {
     'index',
     'onClick',
     'pointing',
+    'shape',
     'type',
   ]