+++ src/components/Menu/menuItemRules.ts
@@ -1,64 +1,84 @@
 import { pxToRem } from '../../lib'
 
+const underlinedItem = (color: string) => ({
+  borderBottom: `solid 5px ${color}`,
+  transition: 'color .1s ease',
+})
+
 export default {
-  root: ({ props: { active, pointing, type } }) => ({
-    color: 'rgba(0, 0, 0, .87)',
-    lineHeight: 1,
-    position: 'relative',
-    verticalAlign: 'middle',
-    padding: `${pxToRem(14)} ${pxToRem(18)}`,
-    cursor: 'pointer',
-    display: 'block',
-    ...(type === 'secondary' && {
-      ...(!pointing && {
+  root: ({ props, variables }) => {
+    const { active, shape, type } = props
+    return {
+      color: variables.defaultColor,
+      lineHeight: 1,
+      position: 'relative',
+      verticalAlign: 'middle',
+      padding: `${pxToRem(14)} ${pxToRem(18)}`,
+      cursor: 'pointer',
+      display: 'block',
+      ...(shape === 'pills' && {
         margin: `0 ${pxToRem(8)} 0 0`,
         borderRadius: pxToRem(5),
       }),
-      ...(pointing && {
+      ...(shape === 'underlined' && {
         margin: '0',
         background: 'transparent',
-        borderColor: '#1b1c1d',
         boxShadow: 'none',
-        color: 'rgba(0,0,0,.95)',
+        color: variables.defaultColor,
       }),
-    }),
-
-    ...(type === 'primary ' && {
-      ':before': {
-        position: 'absolute',
-        content: '""',
-        top: 0,
-        right: 0,
-        height: '100%',
-        width: '1px',
-        background: 'rgba(34, 36, 38, .1)',
-      },
-    }),
-
-    ':hover': {
-      // all menus should have gray background on hover except the secondary pointing menu
-      ...(!(pointing && type === 'secondary') && {
-        background: 'rgba(0, 0, 0, .03)',
+      ...((!shape || shape === 'pointing') && {
+        ':before': {
+          position: 'absolute',
+          content: '""',
+          top: 0,
+          right: 0,
+          height: '100%',
+          width: '1px',
+          background: variables.defaultBorderColor,
+          ...(type === 'primary' && {
+            background: variables.typePrimaryBorderColor,
+          }),
+        },
       }),
-      color: 'rgba(0, 0, 0, .95)',
-    },
 
-    ...(active && {
-      ...(!(pointing && type === 'secondary') && {
-        background: 'rgba(0, 0, 0, .05)',
-      }),
-      color: 'rgba(0, 0, 0, .95)',
       ':hover': {
-        ...(!(pointing && type === 'secondary') && {
-          background: 'rgba(0, 0, 0, .05)',
+        color: variables.defaultActiveColor,
+        // all menus should have gray background on hover except the underlined menu
+        ...(shape !== 'underlined' && {
+          background: variables.defaultActiveBackgroundColor,
+          ...(type === 'primary' && {
+            background: variables.typePrimaryActiveBackgroundColor,
+          }),
         }),
-        color: 'rgba(0, 0, 0, .95)',
+        ...(shape === 'underlined' && {
+          ...underlinedItem(variables.defaultActiveBackgroundColor),
+          ...(type === 'primary' && {
+            ...underlinedItem(variables.typePrimaryActiveBorderColor),
+          }),
+        }),
       },
-      ...(type === 'primary' && {
-        ...(pointing && {
+
+      ...(active && {
+        ...(shape !== 'underlined' && {
+          background: variables.defaultActiveBackgroundColor,
+          ...(type === 'primary' && {
+            background: variables.typePrimaryActiveBackgroundColor,
+          }),
+        }),
+        color: variables.defaultColor,
+        ':hover': {
+          ...(shape !== 'underlined' && {
+            color: variables.defaultActiveColor,
+            background: variables.defaultActiveBackgroundColor,
+            ...(type === 'primary' && {
+              background: variables.typePrimaryActiveBackgroundColor,
+            }),
+          }),
+        },
+        ...(shape === 'pointing' && {
           ':after': {
             visibility: 'visible',
-            background: '#f2f2f2',
+            background: variables.defaultActiveBackgroundColor,
             position: 'absolute',
             content: '""',
             top: '100%',
@@ -68,24 +88,29 @@ export default {
             width: pxToRem(10),
             height: pxToRem(10),
             border: 'none',
-            borderBottom: '1px solid #d4d4d5',
-            borderRight: '1px solid #d4d4d5',
+            borderBottom: `1px solid ${variables.defaultBorderColor}`,
+            borderRight: `1px solid ${variables.defaultBorderColor}`,
             zIndex: '2',
             transition: 'background .1s ease',
+            ...(type === 'primary' && {
+              background: variables.typePrimaryActiveBackgroundColor,
+              borderBottom: `1px solid ${variables.typePrimaryBorderColor}`,
+              borderRight: `1px solid ${variables.typePrimaryBorderColor}`,
+            }),
           },
         }),
-      }),
-      ...(type === 'secondary' &&
-        pointing && {
+        ...(shape === 'underlined' && {
+          color: variables.defaultColor,
           fontWeight: '700',
-          borderColor: '#1b1c1d',
-          borderBottom: 'solid 2px',
-          marginBottom: '-2px',
-          transition: 'color .1s ease',
+          ...underlinedItem(variables.defaultActiveColor),
+          ...(type === 'primary' && {
+            color: variables.typePrimaryActiveColor,
+            ...underlinedItem(variables.typePrimaryActiveColor),
+          }),
         }),
-    }),
-  }),
-
+      }),
+    }
+  },
   anchor: () => ({
     color: 'inherit',
     ':hover': {