+++ src/components/Menu/menuRules.ts
@@ -1,19 +1,29 @@
 import { pxToRem } from '../../lib'
 
+const solidBorder = (color: string) => ({
+  border: `1px solid ${color}`,
+})
+
 export default {
-  root: ({ props: { pointing, type } }) => ({
-    display: 'flex',
-    ...(type === 'primary' && {
-      border: '1px solid rgba(34, 36, 38, .15)',
-      borderRadius: pxToRem(4),
-    }),
-    ...(type === 'secondary' &&
-      pointing && {
-        borderBottom: '2px solid rgba(34, 36, 38, .15)',
+  root: ({ props, variables }) => {
+    const { type, shape } = props
+    return {
+      display: 'flex',
+      ...(shape !== 'pills' &&
+        shape !== 'underlined' && {
+          ...solidBorder(variables.defaultBorderColor),
+          ...(type === 'primary' && {
+            ...solidBorder(variables.typePrimaryBorderColor),
+          }),
+          borderRadius: pxToRem(4),
+        }),
+      ...(shape === 'underlined' && {
+        borderBottom: `1px solid ${variables.defaultBorderColor}`,
       }),
-    minHeight: pxToRem(28),
-    margin: 0,
-    padding: 0,
-    listStyleType: 'none',
-  }),
+      minHeight: pxToRem(28),
+      margin: 0,
+      padding: 0,
+      listStyleType: 'none',
+    }
+  },
 }