+++ src/themes/teams/siteVariables.ts
@@ -8,6 +8,7 @@ export const htmlFontSize = '14px' // what 1rem represents
 
 const blackRgbaFormat = alpha => `rgba(37, 36, 36, ${alpha})`
 export const black = blackRgbaFormat(1)
+export const gray01 = blackRgbaFormat(0.95)
 export const gray02 = blackRgbaFormat(0.75)
 export const gray03 = blackRgbaFormat(0.65)
 export const gray04 = blackRgbaFormat(0.5)